Update GBA.emu to VBA-M GIT 967426e (2025.08.20)
* Imagine: Add Buffer::reset()
* GBA.emu: Handle ROMs over 32MB
* GBA.emu: Handle the multiboot .mb extension
* GBA.emu: Add better error checks for ROM size

+++ GBA.emu/src/core/base/patch.cpp
@@ -286,8 +286,8 @@ bool patchApplyUPS(FILE* f, uint8_t **rom, int *size)
     return false;
   }
   if (dataSize > *size) {
-   // *rom = (uint8_t*)realloc(*rom, dataSize);
-    memset(*rom + *size, 0, dataSize - *size);
+   // *rom = (uint8_t*)realloc(*rom, (size_t)dataSize);
+    memset(*rom + *size, 0, (size_t)(dataSize - *size));
     *size = (int)(dataSize);
   }
